refactor(api): extract media domain (#35)
